requirements.txt
pytest~=8.3.5
requests~=2.32.3
python-dotenv~=1.1.0
pytest-mock
pytest-html
pytest-cov